Split into '__init__' and 'models' files
